+++ anaf/core/administration/forms.py
@@ -169,16 +169,13 @@ class SettingsForm(forms.Form):
                     ModuleSetting.set_for_module(
                         'logopath', logopath, 'anaf.core')
 
-                elif isinstance(self.fields['logo'], forms.ChoiceField):
-                    if self.cleaned_data['logo'] == 'delete':
-                        try:
-                            ModuleSetting.get_for_module(
-                                'anaf.core', 'logopath').delete()
-                        except:
-                            pass
-
+                elif isinstance(self.fields['logo'], forms.ChoiceField) and self.cleaned_data['logo'] == 'delete':
+                    try:
+                        ModuleSetting.get_for_module(
+                            'anaf.core', 'logopath').delete()
+                    except:
+                        pass
             return True
-
         except:
             return False
